Updated from JGit ---------------------------------------------------- JGit commit cefd59f98bfe90c0000ec6a2e515f8d536907abc Merge "Revert Ie48d6556" ---------------------------------------------------- JGit commit fbd61646b3f82bb199bb2e2f1830feb64d84e1af Merge "Proper handling of rebase during pull" ---------------------------------------------------- JGit commit 1ef0343a0ce989d2c77859fe2207d09ff93f6b2b Revert Ie48d6556
This was a mistake that was missed due to historical reasons.

"The first /r/ tells our Apache to redirect the request to Gerrit.
The second /r/ tells Gerrit that the thing following is a Git SHA-1
and it should try to locate the changes that use that commit object.
Nothing I can easily do about it now. The second /r/ is historical
and comes from Gerrit 1.x days."

Change-Id: Iec2dbf5e077f29c0e0686cab11ef197ffc705012
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>

----------------------------------------------------
JGit commit b15b9d5df25b375108ca125b31338df637454865
Proper handling of rebase during pull

After consulting with Christian Halstrick, it turned out that the
handling of rebase during pull was implemented incorrectly.

Change-Id: I40f03409e080cdfeceb21460150f5e02a016e7f4
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>