Updated from JGit ---------------------------------------------------- JGit commit c13bf057540f11e2503e1205d46c800f9d00fd7b Merge "Fix NullPointer when pulling from a deleted local branch" ---------------------------------------------------- JGit commit 68459b646e6c8243dc5650cbe08ce1bcdb8a621d Fix NullPointer when pulling from a deleted local branch
A checked Exception is thrown instead.

The reason for throwing an Exception is that the state of the
repository is inconsistent in this case: There is a merge
configuration containing a non-existing local branch. Ideally the
deletion of a local branch should also delete the corresponding
merge configuration.

Bug: 337315
Change-Id: I71e56ffb90e11e6e3c1bbd964ad63972d67990c0
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
----------------------------------------------------
JGit commit bd531eb998e233b8be27ef9944cae2c1148db5b6
smart-http: Support progress in ReceivePack

As PackParser supports a progress meter for the "Resolving deltas"
phase of its work, we should export this to smart HTTP clients so
they know the server is still working on their (large) upload.

However this isn't as simple as just dropping in a binding for
the SmartOutputStream to flush when its told to.  We want to
avoid spurious flushes triggered by the use of sideband, or the
status report formatting in the send-pack/receive-pack protocol.

Change-Id: Ibd88022a298c5fed0edb23dfaf2e90278807ba8b
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>