Fixed some stream issues.
When converting an Input/OutputStream to Sys.IO.Stream, always create
a new wrapper, instead of reusing an existing Stream object. This is
necessary because the Input/OutputStream may have overriden methods
which need to be called when used through the .net Stream. Also
provided a basic implementation of Mark/Reset for WrappedSystemStream.