[NSch] Allow ssh keys with public exponents which are not 35
We had hard coded '35' as the public exponent which made things fail
for every private key which did not use 35 as the public exponent.
I do not know how this is supposed to be handled, but this is the
only way I can sanely supply the public exponent when signing data.
This fixes https://bugzilla.xamarin.com/show_bug.cgi?id=356.

If there is a better way to fix this, I'm all ears. As it is i 'fixed'
it by modifying our declarations of the java interfaces in order to
pass in the required parameter which just seems wrong.