[Sharpen] Properly handle when a file collides with a directory under linux
Modify all places where we try to get a UnixFileSystemInfo to handle the case
where we probe for the file/directory /foo/bar/baz but /foo/bar is actually
a file. This throws a DirectoryNotFound exception when jgit expects things
to work as if the file/directory did not exist.

Fixes some jgit unit tests.