Modify the test slightly so it passes.
We call enumerator.Next when instantiating the EnumeratorWrapper. This
is so we have a valid value to return if the user checks the 'HasNext'
property. Unfortunately this triggers the exception to be thrown before
the test has entered the try/catch block and makes it fail. Instead of
fixing this properly, just instantiate the enumerator in the try/catch
as the overall effect is the same.