Update ScheduledThreadPoolExecutor.cs
There is a small chance that -1 could be a calculated result for the next time to wait, which would result in an inifinite wait duration (instead of 0).  Since AddTask will not Set if the new task isnt added in the first slot, this has the result of completely stalling the scheduler.